requirements.txt
inflect
mysql-connector-python
terminaltables
jinja2
werkzeug
